pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "promptfoundry"
version = "0.1.0"
description = "Optimization-driven prompt engineering tool"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "PromptFoundry Contributors" }
]
keywords = [
    "llm",
    "prompt-engineering",
    "optimization",
    "evolutionary-algorithms",
    "bayesian-optimization",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]

dependencies = [
    "pydantic>=2.0,<3.0",
    "httpx>=0.24,<1.0",
    "rich>=13.0,<14.0",
    "typer>=0.9,<1.0",
    "pyyaml>=6.0,<7.0",
    "numpy>=1.24,<2.0",
]

[project.optional-dependencies]
bayesian = [
    "optuna>=3.0,<4.0",
]
evolutionary = [
    "deap>=1.4,<2.0",
]
full = [
    "promptfoundry[bayesian,evolutionary]",
]
dev = [
    "pytest>=7.0,<8.0",
    "pytest-cov>=4.0,<5.0",
    "pytest-asyncio>=0.21,<1.0",
    "mypy>=1.5,<2.0",
    "ruff>=0.1,<1.0",
    "pre-commit>=3.0,<4.0",
    "types-PyYAML>=6.0",
    "respx>=0.20,<1.0",
]

[project.scripts]
promptfoundry = "promptfoundry.cli:app"

[project.urls]
Homepage = "https://github.com/IMAGINARY9/PromptFoundry"
Documentation = "https://github.com/IMAGINARY9/PromptFoundry/tree/main/docs"
Repository = "https://github.com/IMAGINARY9/PromptFoundry"
Issues = "https://github.com/IMAGINARY9/PromptFoundry/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/promptfoundry"]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "RUF",    # Ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["promptfoundry"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_configs = true
show_error_codes = true

[[tool.mypy.overrides]]
module = [
    "deap.*",
    "optuna.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-ra -q"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests requiring external services",
]

[tool.coverage.run]
source = ["src/promptfoundry"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
    "@abstractmethod",
]
